Reduce sizes of Monster (goal)Vars
sizeof(Monster): 168 -> 144

+++ Source/monster.cpp
@@ -1945,7 +1945,7 @@ Direction Turn(Direction direction, bool turnLeft)
 	return turnLeft ? Left(direction) : Right(direction);
 }
 
-bool RoundWalk(int monsterId, Direction direction, int *dir)
+bool RoundWalk(int monsterId, Direction direction, int8_t *dir)
 {
 	Monster &monster = Monsters[monsterId];
